refactor(tools): extract registerEditingTools into tools/editing.js
Phase C item 6 (final): moves propose_edit, commit_edit, discard_edit,
snapshot_scene, and list_snapshots out of index.js. pendingProposals and
generateProposalId are passed through toolContext so the shared proposal
state remains a single module-level Map (also referenced by
describe_workflows). Removes z, getFileWriteDiagnostics, readMeta,
indexSceneFile, createSnapshot, and listSnapshots from index.js imports.
